Merge remote-tracking branch 'origin/master' into gehn_monorepo_merge

Conflicts:
	README.md
	Scripts/Python/ChisoPreniv.py
	Scripts/Python/GoMePubNew.py
	Scripts/Python/VeeTsah.py
	Scripts/Python/xCalendarStar.py
	Scripts/Python/xLinkingBookDefs.py
	Scripts/SDL/ChisoPreniv.sdl
	Scripts/SDL/GoMePubNew.sdl
	Scripts/SDL/GreatTreePub.sdl
	Scripts/SDL/VeeTsah.sdl
	Scripts/dat/ChisoPreniv.age
	Scripts/dat/ChisoPreniv.fni
	Scripts/dat/GoMePubNew.age
	Scripts/dat/GoMePubNew.fni
	Scripts/dat/VeeTsah.age
	Scripts/dat/VeeTsah.fni
	Sources/Plasma/Apps/plClient/external/CMakeLists.txt
	Sources/Plasma/Apps/plClient/external/render_svg.py
	Sources/Plasma/FeatureLib/pfConsole/pfConsole.cpp
	Sources/Plasma/FeatureLib/pfPasswordStore/pfPasswordStore.cpp
	Sources/Plasma/FeatureLib/pfPatcher/pfPatcher.cpp
	Sources/Plasma/FeatureLib/pfPython/pyVaultNode.cpp
	Sources/Plasma/FeatureLib/pfPython/pyVaultNodeRef.cpp
	Sources/Plasma/FeatureLib/pfPython/pyVaultPlayerInfoListNode.cpp
	Sources/Plasma/PubUtilLib/plNetGameLib/Private/plNglAuth.cpp
	Sources/Plasma/PubUtilLib/plVault/plVaultClientApi.cpp

+++ Sources/Plasma/NucleusLib/inc/plDrawable.h
@@ -44,7 +44,6 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 
 #include "pnKeyedObject/hsKeyedObject.h"
 #include "plLoadMask.h"
-#include "hsTemplates.h"
 #include "plRenderLevel.h"
 
 class plSpaceTree;
@@ -176,8 +175,8 @@ public:
     virtual void            SetDISpanVisSet(uint32_t diIndex, hsKeyedObject* reg, bool on) = 0;
 
     // Taking span index. DI Index doesn't make sense here, because one object's DI can dereference into many materials etc.
-    virtual hsGMaterial*    GetSubMaterial(int index) const = 0;
-    virtual bool            GetSubVisDists(int index, float& minDist, float& maxDist) const = 0; // return true if span invisible before minDist and/or after maxDist
+    virtual hsGMaterial*    GetSubMaterial(size_t index) const = 0;
+    virtual bool            GetSubVisDists(size_t index, float& minDist, float& maxDist) const = 0; // return true if span invisible before minDist and/or after maxDist
 
     // Should implement hsKeyedObject Read/Write/Save/Load as well
